requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15
django-cors-headers>=4.4
whitenoise>=6.7
cloudinary>=1.41
social-auth-app-django>=5.7
channels>=4.3
daphne>=4.2
channels-redis>=4.2
cryptography>=46.0
